+++ src/utils/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.4.0 (provisional)
+
+- Dropping hidden internals.
+- Adapting `renameGraphKeys` & `updateGraphKeys` to latest graphology version.
+
 ## 2.3.2
 
 - Adding missing `getters.js` file to npm package.